+++ tp1/scr/backtracking/Backtracking.h
@@ -12,12 +12,12 @@ class Backtracking : public DesitionTreeStrategy {
 public:
     bool
     strategyOptimization(int requestsIndex, std::vector<int> *electionTree, std::vector<Request> *requests,
-                             int capacity, int actualMaximum) override;
+                         double capacity, int actualMaximum) override;
 
     int sumSelectedRequestsCosts(int index, std::vector<int> *electionTree, std::vector<Request> *requests);
 
     bool factabilityCut(int requestsIndex, std::vector<int> *electionTree, std::vector<Request> *requests,
-                        int capacity);
+                        double capacity);
 
     int sumSelectedRequestsBenefits(int requestsIndex, std::vector<int> *electionTree, std::vector<Request> *requests);
