+++ tp1/scr/meet_in_the_middle/MeetInTheMiddle.cpp
@@ -8,7 +8,7 @@
 #include "../Solution.h"
 #include <algorithm>
 
-int MeetInTheMiddle::maximumBenefit(int capacity, std::vector <Request> *requests) {
+int MeetInTheMiddle::maximumBenefit(double capacity, std::vector<Request> *requests) {
     KnapsackDesitionTree *bruteForce = new KnapsackDesitionTree(new BruteForce());
     std::vector<Request>::iterator last = requests->begin() + (requests->size()/2);
 
@@ -26,8 +26,9 @@ int MeetInTheMiddle::maximumBenefit(int capacity, std::vector <Request> *request
     return mergeSolutions(&firstHalfSolutions,&secondHalfSolutions, capacity);
 }
 
-int MeetInTheMiddle::mergeSolutions(std::vector<Solution> *firstHalfSolutions, std::vector<Solution> *secondHalfSolutions,
-                                int capacity) {
+int MeetInTheMiddle::mergeSolutions(std::vector<Solution> *firstHalfSolutions,
+                                    std::vector<Solution> *secondHalfSolutions,
+                                    double capacity) {
     std::sort(firstHalfSolutions->begin(),firstHalfSolutions->end(), solutionCompare);
     std::sort(secondHalfSolutions->begin(),secondHalfSolutions->end(), solutionCompare);
